WIP GUI idea for managing web stuff
