+++ src/main/java/net/pms/dlna/DLNAMediaDatabase.java
@@ -44,7 +44,7 @@ import org.h2.jdbcx.JdbcDataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import net.pms.newgui.NavigationShareTab;
+import net.pms.newgui.SharedContentTab;
 
 /**
  * This class provides methods for creating and maintaining the database where
@@ -1378,7 +1378,7 @@ public class DLNAMediaDatabase implements Runnable {
 			scanner = new Thread(this, "Library Scanner");
 			scanner.setPriority(scanner.MIN_PRIORITY);
 			scanner.start();
-			NavigationShareTab.setScanLibraryBusy();
+			SharedContentTab.setScanLibraryBusy();
 		}
 	}
