+++ src/main/java/net/pms/newgui/NavigationShareTab.java
@@ -25,27 +25,15 @@ import com.jgoodies.forms.layout.FormLayout;
 import com.sun.jna.Platform;
 import java.awt.Component;
 import java.awt.ComponentOrientation;
-import java.awt.Dimension;
 import java.awt.Font;
-import java.awt.FontMetrics;
-import java.awt.Point;
 import java.awt.event.*;
-import java.io.File;
-import java.util.Vector;
 import javax.swing.*;
-import javax.swing.table.DefaultTableCellRenderer;
-import javax.swing.table.DefaultTableModel;
-import javax.swing.table.TableColumn;
 import net.pms.Messages;
 import net.pms.PMS;
 import net.pms.configuration.PmsConfiguration;
-import net.pms.database.TableFilesStatus;
-import net.pms.dlna.DLNAMediaDatabase;
 import net.pms.newgui.components.AnimatedIcon;
-import net.pms.newgui.components.AnimatedIcon.AnimatedIconFrame;
 import net.pms.newgui.components.CustomJButton;
 import net.pms.newgui.components.JAnimatedButton;
-import net.pms.newgui.components.JImageButton;
 import net.pms.util.CoverSupplier;
 import net.pms.util.FormLayoutUtil;
 import net.pms.util.FullyPlayedAction;
@@ -58,8 +46,6 @@ public class NavigationShareTab {
 	private static final Logger LOGGER = LoggerFactory.getLogger(NavigationShareTab.class);
 	public static final String ALL_DRIVES = Messages.getString("FoldTab.0");
 
-	private JTable FList;
-	private SharedFoldersTableModel folderTableModel;
 	private JCheckBox hideextensions;
 	private JCheckBox hideemptyfolders;
 	private JCheckBox hideengines;
@@ -83,15 +69,10 @@ public class NavigationShareTab {
 	private JCheckBox prettifyfilenames;
 	private JCheckBox episodeTitles;
 	private JCheckBox resume;
-	private JCheckBox isScanSharedFoldersOnStartup;
 	private JComboBox<String> fullyPlayedAction;
 	private JTextField fullyPlayedOutputDirectory;
 	private CustomJButton selectFullyPlayedOutputDirectory;
 	private static final JAnimatedButton scanButton = new JAnimatedButton("button-scan.png");
-	private final AnimatedIcon scanNormalIcon = (AnimatedIcon) scanButton.getIcon();
-	private final AnimatedIcon scanRolloverIcon = (AnimatedIcon) scanButton.getRolloverIcon();
-	private final AnimatedIcon scanPressedIcon = (AnimatedIcon) scanButton.getPressedIcon();
-	private final AnimatedIcon scanDisabledIcon = (AnimatedIcon) scanButton.getDisabledIcon();
 	private static final AnimatedIcon scanBusyIcon = new AnimatedIcon(scanButton, "button-scan-busy.png");
 	private static final AnimatedIcon scanBusyRolloverIcon = new AnimatedIcon(scanButton, "button-cancel.png");
 	private static final AnimatedIcon scanBusyPressedIcon = new AnimatedIcon(scanButton, "button-cancel_pressed.png");
@@ -105,51 +86,12 @@ public class NavigationShareTab {
 	private JCheckBox isShowFolderRecentlyPlayed;
 	private JCheckBox isShowFolderLiveSubtitles;
 
-	public SharedFoldersTableModel getDf() {
-		return folderTableModel;
-	}
-
 	private final PmsConfiguration configuration;
 	private LooksFrame looksFrame;
 
 	NavigationShareTab(PmsConfiguration configuration, LooksFrame looksFrame) {
 		this.configuration = configuration;
 		this.looksFrame = looksFrame;
-	}
-
-	private void updateModel() {
-		if (folderTableModel.getRowCount() == 1 && folderTableModel.getValueAt(0, 0).equals(ALL_DRIVES)) {
-			configuration.setFolders("");
-		} else {
-			StringBuilder folders = new StringBuilder();
-			StringBuilder foldersMonitored = new StringBuilder();
-
-			int i2 = 0;
-			for (int i = 0; i < folderTableModel.getRowCount(); i++) {
-				if (i > 0) {
-					folders.append(',');
-				}
-
-				String directory = (String) folderTableModel.getValueAt(i, 0);
-				boolean monitored = (boolean) folderTableModel.getValueAt(i, 1);
-
-				// escape embedded commas. note: backslashing isn't safe as it conflicts with
-				// Windows path separators:
-				// http://ps3mediaserver.org/forum/viewtopic.php?f=14&t=8883&start=250#p43520
-				folders.append(directory.replace(",", "&comma;"));
-				if (monitored) {
-					if (i2 > 0) {
-						foldersMonitored.append(',');
-					}
-					i2++;
-
-					foldersMonitored.append(directory.replace(",", "&comma;"));
-				}
-			}
-
-			configuration.setFolders(folders.toString());
-			configuration.setFoldersMonitored(foldersMonitored.toString());
-		}
 	}
 
 	private static final String PANEL_COL_SPEC = "left:pref,          50dlu,                pref, 150dlu,                       pref, 25dlu,               pref, 9dlu, pref, default:grow, pref, 25dlu";
@@ -186,8 +128,6 @@ public class NavigationShareTab {
 		+ "p,"                            //
 		+ "9dlu,"                         //
 		+ "fill:default:grow";            // Shared folders
-	private static final String SHARED_FOLDER_COL_SPEC = "left:pref, left:pref, pref, pref, pref, 10dlu, 0:grow";
-	private static final String SHARED_FOLDER_ROW_SPEC = "p, 3dlu, p, 3dlu, fill:default:grow";
 
 	public JComponent build() {
 		// Apply the orientation for the locale
@@ -204,7 +144,6 @@ public class NavigationShareTab {
 
 		// Init all gui components
 		initSimpleComponents(cc);
-		PanelBuilder builderSharedFolder = initSharedFoldersGuiComponents(cc);
 
 		// Build gui with initialized components
 		if (!configuration.isHideAdvancedOptions()) {
@@ -269,13 +208,7 @@ public class NavigationShareTab {
 			builder.add(fullyPlayedAction,                                         FormLayoutUtil.flip(cc.xyw(4,  29, 3), colSpec, orientation));
 			builder.add(fullyPlayedOutputDirectory,                                FormLayoutUtil.flip(cc.xyw(9,  29, 2), colSpec, orientation));
 			builder.add(selectFullyPlayedOutputDirectory,                          FormLayoutUtil.flip(cc.xyw(11, 29, 2), colSpec, orientation));
-
-			builder.add(builderSharedFolder.getPanel(),                             FormLayoutUtil.flip(cc.xyw(1, 31, 12), colSpec, orientation));
-		} else {
-			builder.add(builderSharedFolder.getPanel(), FormLayoutUtil.flip(cc.xyw(1, 1, 12), colSpec, orientation));
 		}
-
-		builder.add(builderSharedFolder.getPanel(), FormLayoutUtil.flip(cc.xyw(1, 31, 12), colSpec, orientation));
 
 		JPanel panel = builder.getPanel();
 
@@ -462,7 +395,7 @@ public class NavigationShareTab {
 			public void itemStateChanged(ItemEvent e) {
 				configuration.setUseCache((e.getStateChange() == ItemEvent.SELECTED));
 				cachereset.setEnabled(configuration.getUseCache());
-				setScanLibraryEnabled(configuration.getUseCache());
+				SharedContentTab.setScanLibraryEnabled(configuration.getUseCache());
 			}
 		});
 
@@ -748,303 +681,5 @@ public class NavigationShareTab {
 			}
 		});
 		selectFullyPlayedOutputDirectory.setEnabled(configuration.getFullyPlayedAction() == FullyPlayedAction.MOVE_FOLDER);
-	}
-
-	private PanelBuilder initSharedFoldersGuiComponents(CellConstraints cc) {
-		// Apply the orientation for the locale
-		ComponentOrientation orientation = ComponentOrientation.getOrientation(PMS.getLocale());
-		String colSpec = FormLayoutUtil.getColSpec(SHARED_FOLDER_COL_SPEC, orientation);
-
-		FormLayout layoutFolders = new FormLayout(colSpec, SHARED_FOLDER_ROW_SPEC);
-		PanelBuilder builderFolder = new PanelBuilder(layoutFolders);
-		builderFolder.opaque(true);
-
-		JComponent cmp = builderFolder.addSeparator(Messages.getString("FoldTab.7"), FormLayoutUtil.flip(cc.xyw(1, 1, 7), colSpec, orientation));
-		cmp = (JComponent) cmp.getComponent(0);
-		cmp.setFont(cmp.getFont().deriveFont(Font.BOLD));
-
-		folderTableModel = new SharedFoldersTableModel();
-		FList = new JTable(folderTableModel);
-		TableColumn column = FList.getColumnModel().getColumn(0);
-		column.setMinWidth(650);
-
-		JPopupMenu popupMenu = new JPopupMenu();
-		JMenuItem menuItemMarkPlayed = new JMenuItem(Messages.getString("FoldTab.75"));
-		JMenuItem menuItemMarkUnplayed = new JMenuItem(Messages.getString("FoldTab.76"));
-
-		menuItemMarkPlayed.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				String path = (String) FList.getValueAt(FList.getSelectedRow(), 0);
-				TableFilesStatus.setDirectoryFullyPlayed(path, true);
-			}
-		});
-
-		menuItemMarkUnplayed.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				String path = (String) FList.getValueAt(FList.getSelectedRow(), 0);
-				TableFilesStatus.setDirectoryFullyPlayed(path, false);
-			}
-		});
-
-		popupMenu.add(menuItemMarkPlayed);
-		popupMenu.add(menuItemMarkUnplayed);
-
-		FList.setComponentPopupMenu(popupMenu);
-		FList.addMouseListener(new TableMouseListener(FList));
-
-		/* An attempt to set the correct row height adjusted for font scaling.
-		 * It sets all rows based on the font size of cell (0, 0). The + 4 is
-		 * to allow 2 pixels above and below the text. */
-		DefaultTableCellRenderer cellRenderer = (DefaultTableCellRenderer) FList.getCellRenderer(0,0);
-		FontMetrics metrics = cellRenderer.getFontMetrics(cellRenderer.getFont());
-		FList.setRowHeight(metrics.getLeading() + metrics.getMaxAscent() + metrics.getMaxDescent() + 4);
-		FList.setIntercellSpacing(new Dimension(8, 2));
-
-		JImageButton but = new JImageButton("button-add-folder.png");
-		but.setToolTipText(Messages.getString("FoldTab.9"));
-		but.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				JFileChooser chooser;
-				try {
-					chooser = new JFileChooser();
-				} catch (Exception ee) {
-					chooser = new JFileChooser(new RestrictedFileSystemView());
-				}
-				chooser.setFileSelectionMode(JFileChooser.DIRECTORIES_ONLY);
-				int returnVal = chooser.showOpenDialog((Component) e.getSource());
-				if (returnVal == JFileChooser.APPROVE_OPTION) {
-					((SharedFoldersTableModel) FList.getModel()).addRow(new Object[]{chooser.getSelectedFile().getAbsolutePath(), true});
-					if (FList.getModel().getValueAt(0, 0).equals(ALL_DRIVES)) {
-						((SharedFoldersTableModel) FList.getModel()).removeRow(0);
-					}
-					updateModel();
-				}
-			}
-		});
-		builderFolder.add(but, FormLayoutUtil.flip(cc.xy(1, 3), colSpec, orientation));
-
-		JImageButton but2 = new JImageButton("button-remove-folder.png");
-		but2.setToolTipText(Messages.getString("FoldTab.36"));
-		but2.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				if (FList.getSelectedRow() > -1) {
-					if (FList.getModel().getRowCount() == 0) {
-						folderTableModel.addRow(new Object[]{ALL_DRIVES, false});
-					} else {
-						PMS.get().getDatabase().removeMediaEntriesInFolder((String) FList.getValueAt(FList.getSelectedRow(), 0));
-					}
-					((SharedFoldersTableModel) FList.getModel()).removeRow(FList.getSelectedRow());
-					updateModel();
-				}
-			}
-		});
-		builderFolder.add(but2, FormLayoutUtil.flip(cc.xy(2, 3), colSpec, orientation));
-
-		JImageButton but3 = new JImageButton("button-arrow-down.png");
-		but3.setToolTipText(Messages.getString("FoldTab.12"));
-		but3.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				for (int i = 0; i < FList.getRowCount() - 1; i++) {
-					if (FList.isRowSelected(i)) {
-						Object  value1 = FList.getValueAt(i, 0);
-						boolean value2 = (boolean) FList.getValueAt(i, 1);
-
-						FList.setValueAt(FList.getValueAt(i + 1, 0), i    , 0);
-						FList.setValueAt(value1                    , i + 1, 0);
-						FList.setValueAt(FList.getValueAt(i + 1, 1), i    , 1);
-						FList.setValueAt(value2                    , i + 1, 1);
-						FList.changeSelection(i + 1, 1, false, false);
-
-						break;
-					}
-				}
-			}
-		});
-		builderFolder.add(but3, FormLayoutUtil.flip(cc.xy(3, 3), colSpec, orientation));
-
-		JImageButton but4 = new JImageButton("button-arrow-up.png");
-		but4.setToolTipText(Messages.getString("FoldTab.12"));
-		but4.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				for (int i = 1; i < FList.getRowCount(); i++) {
-					if (FList.isRowSelected(i)) {
-						Object  value1 = FList.getValueAt(i, 0);
-						boolean value2 = (boolean) FList.getValueAt(i, 1);
-
-						FList.setValueAt(FList.getValueAt(i - 1, 0), i    , 0);
-						FList.setValueAt(value1                    , i - 1, 0);
-						FList.setValueAt(FList.getValueAt(i - 1, 1), i    , 1);
-						FList.setValueAt(value2                    , i - 1, 1);
-						FList.changeSelection(i - 1, 1, false, false);
-
-						break;
-
-					}
-				}
-			}
-		});
-		builderFolder.add(but4, FormLayoutUtil.flip(cc.xy(4, 3), colSpec, orientation));
-
-		scanButton.setToolTipText(Messages.getString("FoldTab.2"));
-		scanBusyIcon.start();
-		scanBusyDisabledIcon.start();
-		scanButton.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				if (configuration.getUseCache()) {
-					DLNAMediaDatabase database = PMS.get().getDatabase();
-
-					if (database != null) {
-						if (database.isScanLibraryRunning()) {
-							int option = JOptionPane.showConfirmDialog(
-								looksFrame,
-								Messages.getString("FoldTab.10"),
-								Messages.getString("Dialog.Question"),
-								JOptionPane.YES_NO_OPTION);
-							if (option == JOptionPane.YES_OPTION) {
-								database.stopScanLibrary();
-								looksFrame.setStatusLine(Messages.getString("FoldTab.41"));
-								scanButton.setEnabled(false);
-								scanButton.setToolTipText(Messages.getString("FoldTab.41"));
-							}
-						} else {
-							database.scanLibrary();
-							scanButton.setIcon(scanBusyIcon);
-							scanButton.setRolloverIcon(scanBusyRolloverIcon);
-							scanButton.setPressedIcon(scanBusyPressedIcon);
-							scanButton.setDisabledIcon(scanBusyDisabledIcon);
-							scanButton.setToolTipText(Messages.getString("FoldTab.40"));
-						}
-					}
-				}
-			}
-		});
-
-		/*
-		 * Hide the scan button in basic mode since it's better to let it be done in
-		 * realtime.
-		 */
-		if (!configuration.isHideAdvancedOptions()) {
-			builderFolder.add(scanButton, FormLayoutUtil.flip(cc.xy(5, 3), colSpec, orientation));
-		}
-
-		scanButton.setEnabled(configuration.getUseCache());
-
-		isScanSharedFoldersOnStartup = new JCheckBox(Messages.getString("NetworkTab.StartupScan"), configuration.isScanSharedFoldersOnStartup());
-		isScanSharedFoldersOnStartup.setToolTipText(Messages.getString("NetworkTab.StartupScanTooltip"));
-		isScanSharedFoldersOnStartup.setContentAreaFilled(false);
-		isScanSharedFoldersOnStartup.addItemListener(new ItemListener() {
-			@Override
-			public void itemStateChanged(ItemEvent e) {
-				configuration.setScanSharedFoldersOnStartup((e.getStateChange() == ItemEvent.SELECTED));
-			}
-		});
-
-		builderFolder.add(isScanSharedFoldersOnStartup, FormLayoutUtil.flip(cc.xy(7, 3), colSpec, orientation));
-
-		File[] folders = PMS.get().getSharedFoldersArray(false);
-		if (folders != null && folders.length > 0) {
-			File[] foldersMonitored = PMS.get().getSharedFoldersArray(true);
-			for (File folder : folders) {
-				boolean isMonitored = false;
-				if (foldersMonitored != null && foldersMonitored.length > 0) {
-					for (File folderMonitored : foldersMonitored) {
-						if (folderMonitored.getAbsolutePath().equals(folder.getAbsolutePath())) {
-							isMonitored = true;
-						}
-					}
-				}
-				folderTableModel.addRow(new Object[]{folder.getAbsolutePath(), isMonitored});
-			}
-		} else {
-			folderTableModel.addRow(new Object[]{ALL_DRIVES, false});
-		}
-
-		JScrollPane pane = new JScrollPane(FList);
-		Dimension d = FList.getPreferredSize();
-		pane.setPreferredSize(new Dimension(d.width, FList.getRowHeight() * 2));
-		builderFolder.add(pane, FormLayoutUtil.flip(cc.xyw(1, 5, 7), colSpec, orientation));
-
-		return builderFolder;
-	}
-
-	public void setScanLibraryEnabled(boolean enabled) {
-		scanButton.setEnabled(enabled);
-		scanButton.setIcon(scanNormalIcon);
-		scanButton.setRolloverIcon(scanRolloverIcon);
-		scanButton.setPressedIcon(scanPressedIcon);
-		scanButton.setDisabledIcon(scanDisabledIcon);
-		scanButton.setToolTipText(Messages.getString("FoldTab.2"));
-	}
-
-	/**
-	 * @todo combine with setScanLibraryEnabled after we are in sync with DMS
-	 */
-	public static void setScanLibraryBusy() {
-		scanButton.setIcon(scanBusyIcon);
-		scanButton.setRolloverIcon(scanBusyRolloverIcon);
-		scanButton.setPressedIcon(scanBusyPressedIcon);
-		scanButton.setDisabledIcon(scanBusyDisabledIcon);
-		scanButton.setToolTipText(Messages.getString("FoldTab.40"));
-	}
-
-	public class SharedFoldersTableModel extends DefaultTableModel {
-		private static final long serialVersionUID = -4247839506937958655L;
-
-		public SharedFoldersTableModel() {
-			super(new String[]{Messages.getString("FoldTab.56"), Messages.getString("FoldTab.65")}, 0);
-		}
-
-		@Override
-		public Class<?> getColumnClass(int columnIndex) {
-			Class clazz = String.class;
-			switch (columnIndex) {
-				case 1:
-					clazz = Boolean.class;
-					break;
-				default:
-					break;
-			}
-			return clazz;
-		}
-
-		@Override
-		public boolean isCellEditable(int row, int column) {
-			return column == 1;
-		}
-
-		@Override
-		public void setValueAt(Object aValue, int row, int column) {
-			Vector rowVector = (Vector) dataVector.elementAt(row);
-			if (aValue instanceof Boolean && column == 1) {
-				rowVector.setElementAt((boolean) aValue, 1);
-			} else {
-				rowVector.setElementAt(aValue, column);
-			}
-			fireTableCellUpdated(row, column);
-			updateModel();
-		}
-	}
-
-	public class TableMouseListener extends MouseAdapter {
-		private JTable table;
-
-		public TableMouseListener(JTable table) {
-			this.table = table;
-		}
-
-		@Override
-		public void mousePressed(MouseEvent event) {
-			// selects the row at which point the mouse is clicked
-			Point point = event.getPoint();
-			int currentRow = table.rowAtPoint(point);
-			table.setRowSelectionInterval(currentRow, currentRow);
-		}
 	}
 }
