+++ src/main/resources/i18n/messages.properties
@@ -244,17 +244,18 @@ LooksFrame.12=Restart server
 LooksFrame.13=The server has to be restarted due to a configuration change
 LooksFrame.18=Status
 LooksFrame.19=Logs
-LooksFrame.20=General Configuration
 LooksFrame.21=Transcoding Settings
-LooksFrame.22=Navigation/Share Settings
 LooksFrame.24=Help
 LooksFrame.25=About
 LooksFrame.26=FOR TESTING ONLY, POSSIBLY UNSTABLE
-LooksFrame.27=Plugin Management
 LooksFrame.28=This restarts the HTTP server, not the application
 LooksFrame.29=Web interface
 LooksFrame.30=This launches our web interface in your default browser
 LooksFrame.BrowserError=An error occurred while trying to launch the default web browser:
+LooksFrame.TabGeneralSettings=General Settings
+LooksFrame.TabNavigationSettings=Navigation Settings
+LooksFrame.TabPlugins=Plugins
+LooksFrame.TabSharedContent=Shared Content
 LooksFrame.URIError=Could not form a valid URL for the web interface. See the log for further details.
 MediaLibrary.98=By Letter
 MediaLibrary.99=By Year
@@ -464,6 +465,8 @@ RendererPanel.8=Specify a file name
 RendererPanel.9=Select a reference file
 RendererPanel.10=Name
 RendererPanel.11=Address
+SharedContentTab.Source=Source
+SharedContentTab.WebContent=Web content
 SleepManager.PreventSleepPlaybackDetails=System idle sleep is prevented during playback
 SleepManager.PreventSleepPlaybackName=Universal Media Server playing
 SleepManager.PreventSleepRunningDetails=System idle sleep is prevented while Universal Media Server is running
